+++ src/routes/board.rs
@@ -100,7 +100,7 @@ pub async fn board_process_form(
         let filtered_msg = data.formatter.filter_tags(trimmed_message).await;
 
         // Checking against the last message (to prevent spam)
-        if let Ok(last_msg) = client.get_nth_most_active(&info.board, 0).await {
+        if let Ok(last_msg) = client.get_last_message(&info.board).await {
             if last_msg.msg == filtered_msg {
                 return web::Redirect::to("/error?error_code=403").see_other();
             }